[Fix - Front] 디렉토리 및 파일 구조 수정 #14
