+++ front/client/src/pages/Rank.tsx
@@ -1,0 +1,12 @@
+import 'styles/rank/Rank.css';
+import RankTable from 'components/rank/RankTable';
+
+function Rank() {
+    return (
+        <div className='rank'>
+            <h1>ranking</h1>
+            <RankTable />
+        </div>
+    )
+}
+export default Rank;
