+++ front/client/src/types/RankTypes.tsx
@@ -1,0 +1,9 @@
+export type userRank = {
+    ranking : [{
+        rank: string;
+        nickName: string;
+        win: number;
+        lose: number;
+        winRate: string;
+    }]
+}
