+++ front/client/tsconfig.json
@@ -18,9 +18,9 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "jsx": "react-jsx"
+    "jsx": "react-jsx",
+    "baseUrl": "src"
+    
   },
-  "include": [
-    "src"
-  ]
+  "include": ["./"]
 }
